rajout des ancrages pour la vue de la page avec revisite de la navbar css en cours car probleme de visibilite
