+++ src/App.css
@@ -7,3 +7,9 @@ hr {
   box-shadow: 1px 1px 1px rgba(0,0,0,0.6);
   width: 100%;
 }
+
+.section-content {
+  max-width: 800px;
+  margin: 0 auto;
+  padding: 40px;
+}
