+++ src/components/pages/Experience.js
@@ -1,0 +1,12 @@
+import React from 'react'
+
+function Experience() {
+  return (
+    <section className="section-content" id="Experiences">
+      <div className="PageTags">Experiences</div>
+			<p className="p-home-section">mon parcours</p>
+    </section>
+  );
+}
+
+export default Experience;
