+++ src/components/shared/NavLink.js
@@ -1,28 +1,107 @@
-import React from 'react'
-import { Link } from 'react-router-dom'
-import '../stylesheet/NavLink.css'
+import React, {Component} from "react";
 
-function NavLink() {
+import { Link, animateScroll as scroll } from "react-scroll";
+
+import "../stylesheet/NavLink.css";
 
 
+class NavLink extends Component {
+  scrollToTop = () => {
+    scroll.scrollToTop();
+  };
+
+  render() {
     return (
-    <header className="header-navlink">
-        <div className="navBar">
-            <ul className={"links"}>
-                <li className="link-li">
-                    <Link className="header-a" to="/">Home</Link>
-                </li>
-                <li className="link-li">
-                    <Link className="header-a" to="/dashboard">Porte Folio</Link>
-                </li>
-                <li className="link-li">
-                    <Link className="header-a" to="/contact">Contact</Link>
-                </li>
-            </ul>
+      <nav className="nav" id="navbar">
+        <div className="nav-content">
+          <ul className="nav-items">
+            <li className="nav-item">
+              <Link
+                activeClass="active"
+                to="Home"
+                spy={true}
+                smooth={true}
+                offset={-70}
+                duration={500}
+              >
+                Home
+              </Link>
+            </li>
+            <li className="nav-item">
+              <Link
+                activeClass="active"
+                to="Presentation"
+                spy={true}
+                smooth={true}
+                offset={-70}
+                duration={500}
+              >
+                Presentation
+              </Link>
+            </li>
+            <li className="nav-item">
+              <Link
+                activeClass="active"
+                to="Experience"
+                spy={true}
+                smooth={true}
+                offset={-70}
+                duration={500}
+              >
+                Experience
+              </Link>
+            </li>
+            <li className="nav-item">
+              <Link
+                activeClass="active"
+                to="PorteFolio"
+                spy={true}
+                smooth={true}
+                offset={-70}
+                duration={500}
+              >
+                Porte Folio
+              </Link>
+            </li>
+            <li className="nav-item">
+              <Link
+                activeClass="active"
+                to="Contact"
+                spy={true}
+                smooth={true}
+                offset={-70}
+                duration={500}
+              >
+                Contact
+              </Link>
+            </li>
+          </ul>
         </div>
-    </header>
+      </nav>
     );
+  }
 }
 
+
+// function NavLink() {
+
+//     return (
+//     <header className="header-navlink">
+//         <div className="navBar">
+//             <ul className={"links"}>
+//                 <li className="link-li">
+//                     <Link activeClass="active" className="header-a" to="/">Home</Link>
+//                 </li>
+//                 <li className="link-li">
+//                     <Link activeClass="active" className="header-a" to="/dashboard">Porte Folio</Link>
+//                 </li>
+//                 <li className="link-li">
+//                     <Link activeClass="active" className="header-a" to="/contact">Contact</Link>
+//                 </li>
+//             </ul>
+//         </div>
+//     </header>
+//     );
+// }
 
 export default NavLink;
